c.2: reminder of build commands, cosmetic changes
Various typos or space formatting.
Remove the gprof output file when cleaning.
Prefer unsigned types when negative values do not make sense.
Prefer bool to int.
Do not assert that a pointer is not NULL right before dereferencing it.
Fix vector_append when the initial capacity is 0 (this does not happen yet).
Check the count after filling a vector.

+++ impls/c.2/linked_list.c
@@ -11,9 +11,9 @@ list list_push(list lst, MalType data_ptr) {
   return new_head;
 }
 
-int list_count(list lst) {
+size_t list_count(list lst) {
 
-  int counter = 0;
+  size_t counter = 0;
 
   while(lst) {
